Implemented basic user permissions and associated user to post with a foreign Key

+++ src/posts/views.py
@@ -10,7 +10,7 @@ except:
 
 from django.contrib import messages
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.shortcuts import render, get_object_or_404, redirect
 
 from .forms import PostForm
@@ -49,9 +49,16 @@ def post_detail(request, slug=None):
     return render(request, 'post_detail.html', context)
 
 def post_create(request):
+    if not request.user.is_staff and not request.user.is_superuser:
+        raise Http404
+
+    # if not request.user.is_authenticated:
+    #     raise Http404
+
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
+        instance.user = request.user
         instance.save()
         messages.success(request, 'Successfully created.')
         return HttpResponseRedirect(instance.get_absolute_url())
@@ -62,6 +69,9 @@ def post_create(request):
     return render(request, 'post_form.html', context)
 
 def post_update(request, slug=None):
+    if not request.user.is_staff and not request.user.is_superuser:
+        raise Http404
+
     instance = get_object_or_404(Post,slug=slug)
     form = PostForm(request.POST or None, request.FILES or None, instance=instance)
     if form.is_valid():
@@ -78,6 +88,9 @@ def post_update(request, slug=None):
     return render(request, 'post_form.html', context)
 
 def post_delete(request, slug=None):
+    if not request.user.is_staff and not request.user.is_superuser:
+        raise Http404
+
     instance = get_object_or_404(Post,slug=slug)
     instance.delete()
     messages.success(request, 'Successfully deleted.')
